// api/src/constants/producer.ts
// eslint-disable-next-line import/prefer-default-export
export enum Producer {
  id = '_id',
  nif = 'nif',
  bi = 'bi',
  county = 'county',
  cooperativeId = 'cooperative',
  isProducer = 'is_producer',
  companyName = 'company_name',
  shortCode = 'short_code',
  userId = 'user',
  CreatedAt = 'createdAt',
  UpdatedAt = 'updatedAt',
}